[SYSTEMML-148] Refactored .parser.{antlr4,python} to .parser.{dml,pydml}.
Closes #44.
