+++ .gitignore
@@ -41,14 +41,14 @@ src/main/java/Dml.tokens
 src/main/java/DmlLexer.tokens
 src/main/java/Pydml.tokens
 src/main/java/PydmlLexer.tokens
-src/main/java/org/apache/sysml/parser/antlr4/DmlBaseListener.java
-src/main/java/org/apache/sysml/parser/antlr4/DmlLexer.java
-src/main/java/org/apache/sysml/parser/antlr4/DmlListener.java
-src/main/java/org/apache/sysml/parser/antlr4/DmlParser.java
-src/main/java/org/apache/sysml/parser/python/PydmlBaseListener.java
-src/main/java/org/apache/sysml/parser/python/PydmlLexer.java
-src/main/java/org/apache/sysml/parser/python/PydmlListener.java
-src/main/java/org/apache/sysml/parser/python/PydmlParser.java
+src/main/java/org/apache/sysml/parser/dml/DmlBaseListener.java
+src/main/java/org/apache/sysml/parser/dml/DmlLexer.java
+src/main/java/org/apache/sysml/parser/dml/DmlListener.java
+src/main/java/org/apache/sysml/parser/dml/DmlParser.java
+src/main/java/org/apache/sysml/parser/pydml/PydmlBaseListener.java
+src/main/java/org/apache/sysml/parser/pydml/PydmlLexer.java
+src/main/java/org/apache/sysml/parser/pydml/PydmlListener.java
+src/main/java/org/apache/sysml/parser/pydml/PydmlParser.java
 src/test/scripts/**/in
 src/test/scripts/**/out
 src/test/scripts/**/expected
