+++ src/main/java/org/apache/sysml/parser/pydml/ExpressionInfo.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.sysml.parser.python;
+package org.apache.sysml.parser.pydml;
 
 /**
  * This class exists solely to prevent compiler warnings.
@@ -27,6 +27,6 @@ package org.apache.sysml.parser.python;
  * (R-like and Python-like dialects), and Antlr-generated code assumes that
  * these classes are present in the parser's namespace.
  */
-class ExpressionInfo extends org.apache.sysml.parser.antlr4.ExpressionInfo {
+class ExpressionInfo extends org.apache.sysml.parser.dml.ExpressionInfo {
 
 }
