+++ src/main/java/org/apache/sysml/parser/pydml/PyDMLParserWrapper.java
@@ -17,29 +17,26 @@
  * under the License.
  */
 
-package org.apache.sysml.parser.antlr4;
+package org.apache.sysml.parser.pydml;
 
-import java.io.BufferedReader;
 import java.io.ByteArrayInputStream;
 import java.io.FileNotFoundException;
-import java.io.FileReader;
 import java.io.IOException;
 import java.io.InputStream;
-import java.io.InputStreamReader;
 import java.util.HashMap;
 import java.util.Map;
 
+import org.antlr.v4.runtime.ANTLRInputStream;
 import org.antlr.v4.runtime.BailErrorStrategy;
+import org.antlr.v4.runtime.CommonTokenStream;
 import org.antlr.v4.runtime.DefaultErrorStrategy;
 import org.antlr.v4.runtime.atn.PredictionMode;
 import org.antlr.v4.runtime.misc.ParseCancellationException;
+import org.antlr.v4.runtime.tree.ParseTree;
 import org.antlr.v4.runtime.tree.ParseTreeWalker;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
-import org.apache.hadoop.fs.FileSystem;
-import org.apache.hadoop.fs.Path;
 import org.apache.sysml.api.DMLScript;
-import org.apache.sysml.conf.ConfigurationManager;
 import org.apache.sysml.parser.AParserWrapper;
 import org.apache.sysml.parser.DMLProgram;
 import org.apache.sysml.parser.ForStatement;
@@ -56,46 +53,28 @@ import org.apache.sysml.parser.Statement;
 import org.apache.sysml.parser.StatementBlock;
 import org.apache.sysml.parser.WhileStatement;
 import org.apache.sysml.parser.WhileStatementBlock;
-import org.apache.sysml.parser.antlr4.DmlParser.DmlprogramContext;
-import org.apache.sysml.parser.antlr4.DmlParser.FunctionStatementContext;
-import org.apache.sysml.parser.antlr4.DmlParser.StatementContext;
-import org.apache.sysml.parser.antlr4.DmlSyntacticErrorListener.CustomDmlErrorListener;
-import org.apache.sysml.runtime.util.LocalFileUtils;
+import org.apache.sysml.parser.dml.DMLParserWrapper;
+import org.apache.sysml.parser.pydml.PydmlParser.FunctionStatementContext;
+import org.apache.sysml.parser.pydml.PydmlParser.PmlprogramContext;
+import org.apache.sysml.parser.pydml.PydmlParser.StatementContext;
+import org.apache.sysml.parser.pydml.PydmlSyntacticErrorListener.CustomDmlErrorListener;
 
 /**
- * This is the main entry point for the Antlr4 parser.
- * Dml.g4 is the grammar file which enforces syntactic structure of DML program. 
- * DmlSyntaticValidator on other hand captures little bit of semantic as well as does the job of translation of Antlr AST to DMLProgram.
- * At a high-level, DmlSyntaticValidator implements call-back methods that are called by walker.walk(validator, tree)
- * The callback methods are of two type: enterSomeASTNode() and exitSomeASTNode()
- * It is important to note that almost every node in AST has either ExpressionInfo or StatementInfo object associated with it.
- * The key design decision is that while "exiting" the node (i.e. callback to exitSomeASTNode), we use information in given AST node and construct an object of type Statement or Expression and put it in StatementInfo or ExpressionInfo respectively. 
- * This way it avoids any bugs due to lookahead and one only has to "think as an AST node", thereby making any changes to parse code much simpler :)
+ * Logic of this wrapper is similar to DMLParserWrapper.
  * 
- * Note: to add additional builtin function, one only needs to modify DmlSyntaticValidator (which is java file and provides full Eclipse tooling support) not g4. 
- * 
- * To separate logic of semantic validation, DmlSyntaticValidatorHelper contains functions that do semantic validation. Currently, there is no semantic validation as most of it is delegated to subsequent validation phase. 
- * 
- * Whenever there is a parse error, it goes through DmlSyntacticErrorListener. This allows us to pipe the error messages to any future pipeline as well as control the format in an elegant manner. 
- * There are three types of messages passed:
- * - Syntactic errors: When passed DML script doesnot conform to syntatic structure enforced by Dml.g4
- * - Validation errors: Errors due to translation of AST to  DMLProgram
- * - Validation warnings: Messages to inform users that there might be potential bug in their program
- * 
- * As of this moment, Antlr4ParserWrapper is stateful and cannot be multithreaded. This is not big deal because each users calls SystemML in different process.
- * If in future we intend to make it multi-threaded, look at cleanUpState method and resolve the dependency accordingly.    
+ * Note: ExpressionInfo and StatementInfo are simply wrapper objects and are reused in both DML and PyDML parsers.
  *
  */
-public class DMLParserWrapper extends AParserWrapper
+public class PyDMLParserWrapper extends AParserWrapper
 {
 	private static final Log LOG = LogFactory.getLog(DMLScript.class.getName());
 
 	/**
-	 * Custom wrapper to convert statement into statement blocks. Called by doParse and in DmlSyntacticValidator for for, parfor, while, ...
+	 * Custom wrapper to convert statement into statement blocks. Called by doParse and in PydmlSyntacticValidator for for, parfor, while, ...
 	 * @param current a statement
 	 * @return corresponding statement block
 	 */
-	public static StatementBlock getStatementBlock(Statement current) {
+	public static StatementBlock getStatementBlock(org.apache.sysml.parser.Statement current) {
 		StatementBlock blk = null;
 		if(current instanceof ParForStatement) {
 			blk = new ParForStatementBlock();
@@ -130,9 +109,7 @@ public class DMLParserWrapper extends AParserWrapper
 	 * @throws ParseException
 	 */
 	@Override
-	public DMLProgram parse(String fileName, String dmlScript, HashMap<String,String> argVals) 
-		throws ParseException 
-	{
+	public DMLProgram parse(String fileName, String dmlScript, HashMap<String,String> argVals) throws ParseException {
 		DMLProgram prog = null;
 		
 		if(dmlScript == null || dmlScript.trim().isEmpty()) {
@@ -140,52 +117,49 @@ public class DMLParserWrapper extends AParserWrapper
 		}
 		
 		// Set the pipeline required for ANTLR parsing
-		DMLParserWrapper parser = new DMLParserWrapper();
+		PyDMLParserWrapper parser = new PyDMLParserWrapper();
 		prog = parser.doParse(fileName, dmlScript, argVals);
 		
 		if(prog == null) {
-			throw new ParseException("One or more errors found during parsing (could not construct AST for file: " + fileName + "). Cannot proceed ahead.");
+			throw new ParseException("One or more errors found during parsing. (could not construct AST for file: " + fileName + "). Cannot proceed ahead.");
 		}
 		return prog;
 	}
 
 	/**
-	 * This function is supposed to be called directly only from DmlSyntacticValidator when it encounters 'import'
+	 * This function is supposed to be called directly only from PydmlSyntacticValidator when it encounters 'import'
 	 * @param fileName
 	 * @return null if atleast one error
 	 */
 	public DMLProgram doParse(String fileName, String dmlScript, HashMap<String,String> argVals) throws ParseException {
 		DMLProgram dmlPgm = null;
 		
-		org.antlr.v4.runtime.ANTLRInputStream in;
+		ANTLRInputStream in;
 		try {
 			if(dmlScript == null) {
-				dmlScript = readDMLScript(fileName);
+				dmlScript = DMLParserWrapper.readDMLScript(fileName);
 			}
 			
 			InputStream stream = new ByteArrayInputStream(dmlScript.getBytes());
 			in = new org.antlr.v4.runtime.ANTLRInputStream(stream);
-//			else {
-//				if(!(new File(fileName)).exists()) {
-//					throw new ParseException("ERROR: Cannot open file:" + fileName);
-//				}
-//				in = new org.antlr.v4.runtime.ANTLRInputStream(new java.io.FileInputStream(fileName));
-//			}
-		} catch (FileNotFoundException e) {
+		} 
+		catch (FileNotFoundException e) {
 			throw new ParseException("ERROR: Cannot find file:" + fileName, e);
-		} catch (IOException e) {
+		} 
+		catch (IOException e) {
 			throw new ParseException("ERROR: Cannot open file:" + fileName, e);
-		} catch (LanguageException e) {
+		} 
+		catch (LanguageException e) {
 			throw new ParseException("ERROR: " + e.getMessage(), e);
 		}
 
-		DmlprogramContext ast = null;
+		PmlprogramContext ast = null;
 		CustomDmlErrorListener errorListener = new CustomDmlErrorListener();
 		
 		try {
-			DmlLexer lexer = new DmlLexer(in);
-			org.antlr.v4.runtime.CommonTokenStream tokens = new org.antlr.v4.runtime.CommonTokenStream(lexer);
-			DmlParser antlr4Parser = new DmlParser(tokens);
+			PydmlLexer lexer = new PydmlLexer(in);
+			CommonTokenStream tokens = new CommonTokenStream(lexer);
+			PydmlParser antlr4Parser = new PydmlParser(tokens);
 			
 			boolean tryOptimizedParsing = false; // For now no optimization, since it is not able to parse integer value. 
 	
@@ -195,7 +169,7 @@ public class DMLParserWrapper extends AParserWrapper
 				antlr4Parser.removeErrorListeners();
 				antlr4Parser.setErrorHandler(new BailErrorStrategy());
 				try{
-					ast = antlr4Parser.dmlprogram();
+					ast = antlr4Parser.pmlprogram();
 					// If successful, no need to try out full LL(*) ... SLL was enough
 				}
 				catch(ParseCancellationException ex) {
@@ -204,17 +178,15 @@ public class DMLParserWrapper extends AParserWrapper
 					antlr4Parser.reset();
 					if(fileName != null) {
 						errorListener.pushCurrentFileName(fileName);
-						// DmlSyntacticErrorListener.currentFileName.push(fileName);
 					}
 					else {
 						errorListener.pushCurrentFileName("MAIN_SCRIPT");
-						// DmlSyntacticErrorListener.currentFileName.push("MAIN_SCRIPT");
 					}
 					// Set our custom error listener
 					antlr4Parser.addErrorListener(errorListener);
 					antlr4Parser.setErrorHandler(new DefaultErrorStrategy());
 					antlr4Parser.getInterpreter().setPredictionMode(PredictionMode.LL);
-					ast = antlr4Parser.dmlprogram();
+					ast = antlr4Parser.pmlprogram();
 				}
 			}
 			else {
@@ -224,7 +196,7 @@ public class DMLParserWrapper extends AParserWrapper
 				errorListener.pushCurrentFileName(fileName);
 	
 				// Now do the parsing
-				ast = antlr4Parser.dmlprogram();
+				ast = antlr4Parser.pmlprogram();
 			}
 		}
 		catch(Exception e) {
@@ -235,11 +207,11 @@ public class DMLParserWrapper extends AParserWrapper
 		try {
 			// Now convert the parse tree into DMLProgram
 			// Do syntactic validation while converting 
-			org.antlr.v4.runtime.tree.ParseTree tree = ast;
+			ParseTree tree = ast;
 			// And also do syntactic validation
-			org.antlr.v4.runtime.tree.ParseTreeWalker walker = new ParseTreeWalker();
-			DmlSyntacticValidatorHelper helper = new DmlSyntacticValidatorHelper(errorListener);
-			DmlSyntacticValidator validator = new DmlSyntacticValidator(helper, errorListener.peekFileName(), argVals);
+			ParseTreeWalker walker = new ParseTreeWalker();
+			PydmlSyntacticValidatorHelper helper = new PydmlSyntacticValidatorHelper(errorListener);
+			PydmlSyntacticValidator validator = new PydmlSyntacticValidator(helper, fileName, argVals);
 			walker.walk(validator, tree);
 			errorListener.popFileName();
 			if(errorListener.isAtleastOneError()) {
@@ -253,8 +225,9 @@ public class DMLParserWrapper extends AParserWrapper
 		
 		return dmlPgm;
 	}
-	
-	private DMLProgram createDMLProgram(DmlprogramContext ast) {
+
+
+	private DMLProgram createDMLProgram(PmlprogramContext ast) {
 
 		DMLProgram dmlPgm = new DMLProgram();
 
@@ -274,10 +247,15 @@ public class DMLParserWrapper extends AParserWrapper
 
 		// Then add all the statements
 		for(StatementContext stmtCtx : ast.blocks) {
-			org.apache.sysml.parser.Statement current = stmtCtx.info.stmt;
+			Statement current = stmtCtx.info.stmt;
 			if(current == null) {
 				LOG.error("line: " + stmtCtx.start.getLine() + ":" + stmtCtx.start.getCharPositionInLine() + " cannot process the statement");
 				return null;
+			}
+			
+			// Ignore Newline logic 
+			if(current.isEmptyNewLineStatement()) {
+				continue;
 			}
 
 			if(current instanceof ImportStatement) {
@@ -286,31 +264,6 @@ public class DMLParserWrapper extends AParserWrapper
 					// Add the DMLProgram entries into current program
 					for(Map.Entry<String, DMLProgram> entry : stmtCtx.info.namespaces.entrySet()) {
 						dmlPgm.getNamespaces().put(entry.getKey(), entry.getValue());
-						
-//						// Don't add DMLProgram into the current program, just add function statements
-						// dmlPgm.getNamespaces().put(entry.getKey(), entry.getValue());
-						// Add function statements to current dml program
-//						DMLProgram importedPgm = entry.getValue();
-//						try {
-//							for(FunctionStatementBlock importedFnBlk : importedPgm.getFunctionStatementBlocks()) {
-//								if(importedFnBlk.getStatements() != null && importedFnBlk.getStatements().size() == 1) {
-//									String functionName = ((FunctionStatement)importedFnBlk.getStatement(0)).getName();
-//									System.out.println("Adding function => " + entry.getKey() + "::" + functionName);
-//									TODO:33
-//									dmlPgm.addFunctionStatementBlock(entry.getKey(), functionName, importedFnBlk);
-//								}
-//								else {
-//									LOG.error("line: " + stmtCtx.start.getLine() + ":" + stmtCtx.start.getCharPositionInLine() + " incorrect number of functions in the imported function block .... strange");
-//									return null;
-//								}
-//							}
-//							if(importedPgm.getStatementBlocks() != null && importedPgm.getStatementBlocks().size() > 0) {
-//								LOG.warn("Only the functions can be imported from the namespace " + entry.getKey());
-//							}
-//						} catch (LanguageException e) {
-//							LOG.error("line: " + stmtCtx.start.getLine() + ":" + stmtCtx.start.getCharPositionInLine() + " cannot import functions from the file in the import statement: " + e.getMessage());
-//							return null;
-//						}
 					}
 				}
 				else {
@@ -326,60 +279,5 @@ public class DMLParserWrapper extends AParserWrapper
 
 		dmlPgm.mergeStatementBlocks();
 		return dmlPgm;
-	}
-	
-	public static String readDMLScript( String script ) 
-			throws IOException, LanguageException
-	{
-		String dmlScriptStr = null;
-		
-		//read DML script from file
-		if(script == null)
-			throw new LanguageException("DML script path was not specified!");
-		
-		StringBuilder sb = new StringBuilder();
-		BufferedReader in = null;
-		try 
-		{
-			//read from hdfs or gpfs file system
-			if(    script.startsWith("hdfs:") 
-				|| script.startsWith("gpfs:") ) 
-			{ 
-				if( !LocalFileUtils.validateExternalFilename(script, true) )
-					throw new LanguageException("Invalid (non-trustworthy) hdfs filename.");
-				FileSystem fs = FileSystem.get(ConfigurationManager.getCachedJobConf());
-				Path scriptPath = new Path(script);
-				in = new BufferedReader(new InputStreamReader(fs.open(scriptPath)));
-			}
-			// from local file system
-			else 
-			{ 
-				if( !LocalFileUtils.validateExternalFilename(script, false) )
-					throw new LanguageException("Invalid (non-trustworthy) local filename.");
-				in = new BufferedReader(new FileReader(script));
-			}
-			
-			//core script reading
-			String tmp = null;
-			while ((tmp = in.readLine()) != null)
-			{
-				sb.append( tmp );
-				sb.append( "\n" );
-			}
-		}
-		catch (IOException ex)
-		{
-			LOG.error("Failed to read the script from the file system", ex);
-			throw ex;
-		}
-		finally 
-		{
-			if( in != null )
-				in.close();
-		}
-		
-		dmlScriptStr = sb.toString();
-		
-		return dmlScriptStr;
 	}
 }
