+++ src/main/java/org/apache/sysml/parser/pydml/PydmlSyntacticValidator.java
@@ -17,10 +17,11 @@
  * under the License.
  */
 
-package org.apache.sysml.parser.antlr4;
+package org.apache.sysml.parser.pydml;
 
 import java.io.File;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -29,70 +30,71 @@ import org.antlr.v4.runtime.ParserRuleContext;
 import org.antlr.v4.runtime.Token;
 import org.antlr.v4.runtime.tree.ErrorNode;
 import org.antlr.v4.runtime.tree.TerminalNode;
+import org.apache.commons.lang.StringUtils;
+import org.apache.sysml.parser.AssignmentStatement;
+import org.apache.sysml.parser.BinaryExpression;
+import org.apache.sysml.parser.BooleanExpression;
+import org.apache.sysml.parser.BooleanIdentifier;
+import org.apache.sysml.parser.BuiltinFunctionExpression;
 import org.apache.sysml.parser.ConditionalPredicate;
+import org.apache.sysml.parser.ConstIdentifier;
 import org.apache.sysml.parser.DMLProgram;
+import org.apache.sysml.parser.DataExpression;
 import org.apache.sysml.parser.DataIdentifier;
 import org.apache.sysml.parser.DoubleIdentifier;
 import org.apache.sysml.parser.Expression;
 import org.apache.sysml.parser.Expression.DataOp;
 import org.apache.sysml.parser.Expression.DataType;
 import org.apache.sysml.parser.Expression.ValueType;
-import org.apache.sysml.parser.antlr4.DmlParser.AddSubExpressionContext;
-import org.apache.sysml.parser.antlr4.DmlParser.AssignmentStatementContext;
-import org.apache.sysml.parser.antlr4.DmlParser.AtomicExpressionContext;
-import org.apache.sysml.parser.antlr4.DmlParser.BooleanAndExpressionContext;
-import org.apache.sysml.parser.antlr4.DmlParser.BooleanNotExpressionContext;
-import org.apache.sysml.parser.antlr4.DmlParser.BooleanOrExpressionContext;
-import org.apache.sysml.parser.antlr4.DmlParser.BuiltinFunctionExpressionContext;
-import org.apache.sysml.parser.antlr4.DmlParser.CommandlineParamExpressionContext;
-import org.apache.sysml.parser.antlr4.DmlParser.CommandlinePositionExpressionContext;
-import org.apache.sysml.parser.antlr4.DmlParser.ConstDoubleIdExpressionContext;
-import org.apache.sysml.parser.antlr4.DmlParser.ConstFalseExpressionContext;
-import org.apache.sysml.parser.antlr4.DmlParser.ConstIntIdExpressionContext;
-import org.apache.sysml.parser.antlr4.DmlParser.ConstStringIdExpressionContext;
-import org.apache.sysml.parser.antlr4.DmlParser.ConstTrueExpressionContext;
-import org.apache.sysml.parser.antlr4.DmlParser.DataIdExpressionContext;
-import org.apache.sysml.parser.antlr4.DmlParser.DataIdentifierContext;
-import org.apache.sysml.parser.antlr4.DmlParser.DmlprogramContext;
-import org.apache.sysml.parser.antlr4.DmlParser.ExpressionContext;
-import org.apache.sysml.parser.antlr4.DmlParser.ExternalFunctionDefExpressionContext;
-import org.apache.sysml.parser.antlr4.DmlParser.ForStatementContext;
-import org.apache.sysml.parser.antlr4.DmlParser.FunctionCallAssignmentStatementContext;
-import org.apache.sysml.parser.antlr4.DmlParser.FunctionCallMultiAssignmentStatementContext;
-import org.apache.sysml.parser.antlr4.DmlParser.FunctionStatementContext;
-import org.apache.sysml.parser.antlr4.DmlParser.IfStatementContext;
-import org.apache.sysml.parser.antlr4.DmlParser.IfdefAssignmentStatementContext;
-import org.apache.sysml.parser.antlr4.DmlParser.ImportStatementContext;
-import org.apache.sysml.parser.antlr4.DmlParser.IndexedExpressionContext;
-import org.apache.sysml.parser.antlr4.DmlParser.InternalFunctionDefExpressionContext;
-import org.apache.sysml.parser.antlr4.DmlParser.IterablePredicateColonExpressionContext;
-import org.apache.sysml.parser.antlr4.DmlParser.IterablePredicateSeqExpressionContext;
-import org.apache.sysml.parser.antlr4.DmlParser.MatrixDataTypeCheckContext;
-import org.apache.sysml.parser.antlr4.DmlParser.MatrixMulExpressionContext;
-import org.apache.sysml.parser.antlr4.DmlParser.Ml_typeContext;
-import org.apache.sysml.parser.antlr4.DmlParser.ModIntDivExpressionContext;
-import org.apache.sysml.parser.antlr4.DmlParser.MultDivExpressionContext;
-import org.apache.sysml.parser.antlr4.DmlParser.ParForStatementContext;
-import org.apache.sysml.parser.antlr4.DmlParser.ParameterizedExpressionContext;
-import org.apache.sysml.parser.antlr4.DmlParser.PathStatementContext;
-import org.apache.sysml.parser.antlr4.DmlParser.PowerExpressionContext;
-import org.apache.sysml.parser.antlr4.DmlParser.RelationalExpressionContext;
-import org.apache.sysml.parser.antlr4.DmlParser.SimpleDataIdentifierExpressionContext;
-import org.apache.sysml.parser.antlr4.DmlParser.StatementContext;
-import org.apache.sysml.parser.antlr4.DmlParser.StrictParameterizedExpressionContext;
-import org.apache.sysml.parser.antlr4.DmlParser.StrictParameterizedKeyValueStringContext;
-import org.apache.sysml.parser.antlr4.DmlParser.TypedArgNoAssignContext;
-import org.apache.sysml.parser.antlr4.DmlParser.UnaryExpressionContext;
-import org.apache.sysml.parser.antlr4.DmlParser.ValueTypeContext;
-import org.apache.sysml.parser.antlr4.DmlParser.WhileStatementContext;
+import org.apache.sysml.parser.pydml.PydmlParser.AddSubExpressionContext;
+import org.apache.sysml.parser.pydml.PydmlParser.AssignmentStatementContext;
+import org.apache.sysml.parser.pydml.PydmlParser.AtomicExpressionContext;
+import org.apache.sysml.parser.pydml.PydmlParser.BooleanAndExpressionContext;
+import org.apache.sysml.parser.pydml.PydmlParser.BooleanNotExpressionContext;
+import org.apache.sysml.parser.pydml.PydmlParser.BooleanOrExpressionContext;
+import org.apache.sysml.parser.pydml.PydmlParser.BuiltinFunctionExpressionContext;
+import org.apache.sysml.parser.pydml.PydmlParser.CommandlineParamExpressionContext;
+import org.apache.sysml.parser.pydml.PydmlParser.CommandlinePositionExpressionContext;
+import org.apache.sysml.parser.pydml.PydmlParser.ConstDoubleIdExpressionContext;
+import org.apache.sysml.parser.pydml.PydmlParser.ConstFalseExpressionContext;
+import org.apache.sysml.parser.pydml.PydmlParser.ConstIntIdExpressionContext;
+import org.apache.sysml.parser.pydml.PydmlParser.ConstStringIdExpressionContext;
+import org.apache.sysml.parser.pydml.PydmlParser.ConstTrueExpressionContext;
+import org.apache.sysml.parser.pydml.PydmlParser.DataIdExpressionContext;
+import org.apache.sysml.parser.pydml.PydmlParser.DataIdentifierContext;
+import org.apache.sysml.parser.pydml.PydmlParser.ExpressionContext;
+import org.apache.sysml.parser.pydml.PydmlParser.ExternalFunctionDefExpressionContext;
+import org.apache.sysml.parser.pydml.PydmlParser.ForStatementContext;
+import org.apache.sysml.parser.pydml.PydmlParser.FunctionCallAssignmentStatementContext;
+import org.apache.sysml.parser.pydml.PydmlParser.FunctionCallMultiAssignmentStatementContext;
+import org.apache.sysml.parser.pydml.PydmlParser.FunctionStatementContext;
+import org.apache.sysml.parser.pydml.PydmlParser.IfStatementContext;
+import org.apache.sysml.parser.pydml.PydmlParser.IfdefAssignmentStatementContext;
+import org.apache.sysml.parser.pydml.PydmlParser.IgnoreNewLineContext;
+import org.apache.sysml.parser.pydml.PydmlParser.ImportStatementContext;
+import org.apache.sysml.parser.pydml.PydmlParser.IndexedExpressionContext;
+import org.apache.sysml.parser.pydml.PydmlParser.InternalFunctionDefExpressionContext;
+import org.apache.sysml.parser.pydml.PydmlParser.IterablePredicateColonExpressionContext;
+import org.apache.sysml.parser.pydml.PydmlParser.IterablePredicateSeqExpressionContext;
+import org.apache.sysml.parser.pydml.PydmlParser.MatrixDataTypeCheckContext;
+import org.apache.sysml.parser.pydml.PydmlParser.Ml_typeContext;
+import org.apache.sysml.parser.pydml.PydmlParser.ModIntDivExpressionContext;
+import org.apache.sysml.parser.pydml.PydmlParser.MultDivExpressionContext;
+import org.apache.sysml.parser.pydml.PydmlParser.ParForStatementContext;
+import org.apache.sysml.parser.pydml.PydmlParser.ParameterizedExpressionContext;
+import org.apache.sysml.parser.pydml.PydmlParser.PathStatementContext;
+import org.apache.sysml.parser.pydml.PydmlParser.PmlprogramContext;
+import org.apache.sysml.parser.pydml.PydmlParser.PowerExpressionContext;
+import org.apache.sysml.parser.pydml.PydmlParser.RelationalExpressionContext;
+import org.apache.sysml.parser.pydml.PydmlParser.SimpleDataIdentifierExpressionContext;
+import org.apache.sysml.parser.pydml.PydmlParser.StatementContext;
+import org.apache.sysml.parser.pydml.PydmlParser.StrictParameterizedExpressionContext;
+import org.apache.sysml.parser.pydml.PydmlParser.StrictParameterizedKeyValueStringContext;
+import org.apache.sysml.parser.pydml.PydmlParser.TypedArgNoAssignContext;
+import org.apache.sysml.parser.pydml.PydmlParser.UnaryExpressionContext;
+import org.apache.sysml.parser.pydml.PydmlParser.ValueDataTypeCheckContext;
+import org.apache.sysml.parser.pydml.PydmlParser.WhileStatementContext;
 import org.apache.sysml.parser.AParserWrapper;
-import org.apache.sysml.parser.AssignmentStatement;
-import org.apache.sysml.parser.BinaryExpression;
-import org.apache.sysml.parser.BooleanExpression;
-import org.apache.sysml.parser.BooleanIdentifier;
-import org.apache.sysml.parser.BuiltinFunctionExpression;
-import org.apache.sysml.parser.ConstIdentifier;
-import org.apache.sysml.parser.DataExpression;
 import org.apache.sysml.parser.ExternalFunctionStatement;
 import org.apache.sysml.parser.ForStatement;
 import org.apache.sysml.parser.FunctionCallIdentifier;
@@ -121,15 +123,15 @@ import org.apache.sysml.parser.WhileStatement;
  * TODO: Refactor duplicated parser code dml/pydml (entire package).
  *
  */
-public class DmlSyntacticValidator implements DmlListener
+public class PydmlSyntacticValidator implements PydmlListener
 {	
-	private DmlSyntacticValidatorHelper helper = null;
-
-	private String _workingDir = ".";   //current working directory 
+	private PydmlSyntacticValidatorHelper helper = null;
+	
+	private String _workingDir = ".";   //current working directory
 	private String _currentPath = null; //current file path
 	private HashMap<String,String> argVals = null;
 	
-	public DmlSyntacticValidator(DmlSyntacticValidatorHelper helper, String currentPath, HashMap<String,String> argVals) {
+	public PydmlSyntacticValidator(PydmlSyntacticValidatorHelper helper, String currentPath, HashMap<String,String> argVals) {
 		this.helper = helper;
 		this.argVals = argVals;
 		
@@ -163,7 +165,12 @@ public class DmlSyntacticValidator implements DmlListener
 	public void enterDataIdExpression(DataIdExpressionContext ctx) { }
 
 	@Override
-	public void enterDmlprogram(DmlprogramContext ctx) { }
+	public void enterIgnoreNewLine(IgnoreNewLineContext ctx) { }
+	@Override
+	public void enterPmlprogram(PmlprogramContext ctx) { }
+	@Override
+	public void exitPmlprogram(PmlprogramContext ctx) { }
+	
 	@Override
 	public void enterEveryRule(ParserRuleContext arg0) {
 		if(arg0 instanceof StatementContext) {
@@ -203,7 +210,6 @@ public class DmlSyntacticValidator implements DmlListener
 	public void enterIndexedExpression(IndexedExpressionContext ctx) { }
 	@Override
 	public void enterInternalFunctionDefExpression(InternalFunctionDefExpressionContext ctx) { }
-	public void enterMatrixMulExpression(MatrixMulExpressionContext ctx) { }
 	@Override
 	public void enterMl_type(Ml_typeContext ctx) { }
 	@Override
@@ -228,8 +234,6 @@ public class DmlSyntacticValidator implements DmlListener
 	public void enterTypedArgNoAssign(TypedArgNoAssignContext ctx) { }
 	@Override
 	public void enterUnaryExpression(UnaryExpressionContext ctx) { }
-	@Override
-	public void enterValueType(ValueTypeContext ctx) { }
 	@Override
 	public void enterWhileStatement(WhileStatementContext ctx) { }
 	
@@ -270,6 +274,7 @@ public class DmlSyntacticValidator implements DmlListener
 	@Override
 	public void exitAddSubExpression(AddSubExpressionContext ctx) {
 		if(ctx.left.info.expr != null && ctx.right.info.expr != null) {
+			// Addition and subtraction operator same as DML
 			Expression.BinaryOp bop = Expression.getBinaryOp(ctx.op.getText());
 			ctx.info.expr = new BinaryExpression(bop);
 			((BinaryExpression)ctx.info.expr).setLeft(ctx.left.info.expr);
@@ -277,10 +282,24 @@ public class DmlSyntacticValidator implements DmlListener
 			setFileLineColumn(ctx.info.expr, ctx);
 		}
 	}
+	
+	
+	
 	@Override
 	public void exitModIntDivExpression(ModIntDivExpressionContext ctx) {
 		if(ctx.left.info.expr != null && ctx.right.info.expr != null) {
-			Expression.BinaryOp bop = Expression.getBinaryOp(ctx.op.getText());
+			String dmlOperator = "";
+			if(ctx.op.getText().compareTo("//") == 0) {
+				dmlOperator = "%/%";
+			}
+			else if(ctx.op.getText().compareTo("%") == 0) {
+				dmlOperator = "%%";
+			}
+			else {
+				helper.notifyErrorListeners("Incorrect operator (expected // or %)", ctx.op);
+				return;
+			}
+			Expression.BinaryOp bop = Expression.getBinaryOp(dmlOperator);
 			ctx.info.expr = new BinaryExpression(bop);
 			((BinaryExpression)ctx.info.expr).setLeft(ctx.left.info.expr);
 			((BinaryExpression)ctx.info.expr).setRight(ctx.right.info.expr);
@@ -330,49 +349,40 @@ public class DmlSyntacticValidator implements DmlListener
 		((BinaryExpression)ctx.info.expr).setRight(ctx.right.info.expr);
 		setFileLineColumn(ctx.info.expr, ctx);
 	}
+	
 	@Override
 	public void exitPowerExpression(PowerExpressionContext ctx) {
-		Expression.BinaryOp bop = Expression.getBinaryOp(ctx.op.getText());
+		String dmlOperator = "";
+		if(ctx.op.getText().compareTo("**") == 0) {
+			dmlOperator = "^";
+		}
+		else {
+			helper.notifyErrorListeners("Incorrect operator (expected **)", ctx.op);
+			return;
+		}
+		
+		Expression.BinaryOp bop = Expression.getBinaryOp(dmlOperator);
 		ctx.info.expr = new BinaryExpression(bop);
 		((BinaryExpression)ctx.info.expr).setLeft(ctx.left.info.expr);
 		((BinaryExpression)ctx.info.expr).setRight(ctx.right.info.expr);
 		setFileLineColumn(ctx.info.expr, ctx);
 	}
 	
-	@Override
-	public void exitMatrixMulExpression(MatrixMulExpressionContext ctx) {
-		Expression.BinaryOp bop = Expression.getBinaryOp(ctx.op.getText());
-		ctx.info.expr = new BinaryExpression(bop);
-		((BinaryExpression)ctx.info.expr).setLeft(ctx.left.info.expr);
-		((BinaryExpression)ctx.info.expr).setRight(ctx.right.info.expr);
-		setFileLineColumn(ctx.info.expr, ctx);
-	}
+	// TODO: 
+//	@Override
+//	public void exitMatrixMulExpression(MatrixMulExpressionContext ctx) {
+//		Expression.BinaryOp bop = Expression.getBinaryOp(ctx.op.getText());
+//		ctx.info.expr = new BinaryExpression(bop);
+//		((BinaryExpression)ctx.info.expr).setLeft(ctx.left.info.expr);
+//		((BinaryExpression)ctx.info.expr).setRight(ctx.right.info.expr);
+//		setFileLineColumn(ctx.info.expr, ctx);
+//	}
 
 	// --------------------------------------------------------------------
 
 	@Override
 	public void exitRelationalExpression(RelationalExpressionContext ctx) {
 		if(ctx.left.info.expr != null && ctx.right.info.expr != null) {
-//			String fileName = helper.getCurrentFileName();
-//			int line = ctx.start.getLine();
-//			int col = ctx.start.getCharPositionInLine();
-//			ArrayList<ParameterExpression> paramExpression = new ArrayList<ParameterExpression>();
-//			paramExpression.add(new ParameterExpression(null, ctx.left.info.expr));
-//			paramExpression.add(new ParameterExpression(null, ctx.right.info.expr));
-//			ParameterExpression operator = new ParameterExpression(null, new StringIdentifier(ctx.op.getText(), fileName, line, col, line, col));
-//			paramExpression.add(operator);
-//			
-//			try {
-//				BuiltinFunctionExpression bife = BuiltinFunctionExpression.getBuiltinFunctionExpression("ppred", paramExpression, fileName, line, col, line, col);
-//				if (bife != null){
-//					// It is a builtin function
-//					ctx.info.expr = bife;
-//					return;
-//				}
-//			}
-//			catch(Exception e) {}
-//			helper.notifyErrorListeners("Cannot parse relational expression", ctx.getStart());
-			
 			Expression.RelationalOp rop = Expression.getRelationalOp(ctx.op.getText());
 			ctx.info.expr = new RelationalExpression(rop);
 			((RelationalExpression)ctx.info.expr).setLeft(ctx.left.info.expr);
@@ -386,7 +396,16 @@ public class DmlSyntacticValidator implements DmlListener
 	@Override
 	public void exitBooleanAndExpression(BooleanAndExpressionContext ctx) {
 		if(ctx.left.info.expr != null && ctx.right.info.expr != null) {
-			Expression.BooleanOp bop = Expression.getBooleanOp(ctx.op.getText());
+			String dmlOperator = "";
+			if(ctx.op.getText().compareTo("&") == 0 || ctx.op.getText().compareTo("and") == 0) {
+				dmlOperator = "&";
+			}
+			else {
+				helper.notifyErrorListeners("Incorrect operator (expected &)", ctx.op);
+				return;
+			}
+			
+			Expression.BooleanOp bop = Expression.getBooleanOp(dmlOperator);
 			ctx.info.expr = new BooleanExpression(bop);
 			((BooleanExpression)ctx.info.expr).setLeft(ctx.left.info.expr);
 			((BooleanExpression)ctx.info.expr).setRight(ctx.right.info.expr);
@@ -397,7 +416,16 @@ public class DmlSyntacticValidator implements DmlListener
 	@Override
 	public void exitBooleanOrExpression(BooleanOrExpressionContext ctx) {
 		if(ctx.left.info.expr != null && ctx.right.info.expr != null) {
-			Expression.BooleanOp bop = Expression.getBooleanOp(ctx.op.getText());
+			String dmlOperator = "";
+			if(ctx.op.getText().compareTo("|") == 0 || ctx.op.getText().compareTo("or") == 0) {
+				dmlOperator = "|";
+			}
+			else {
+				helper.notifyErrorListeners("Incorrect operator (expected |)", ctx.op);
+				return;
+			}
+			
+			Expression.BooleanOp bop = Expression.getBooleanOp(dmlOperator);
 			ctx.info.expr = new BooleanExpression(bop);
 			((BooleanExpression)ctx.info.expr).setLeft(ctx.left.info.expr);
 			((BooleanExpression)ctx.info.expr).setRight(ctx.right.info.expr);
@@ -423,25 +451,6 @@ public class DmlSyntacticValidator implements DmlListener
 		setFileLineColumn(ctx.info.expr, ctx);
 	}
 	
-//	@Override
-//	public void exitConstBooleanIdExpression(ConstBooleanIdExpressionContext ctx) {
-//		boolean val = false;
-//		if(ctx.getText().compareTo("TRUE") == 0) {
-//			val = true;
-//		}
-//		else if(ctx.getText().compareTo("FALSE") == 0) {
-//			val = false;
-//		}
-//		else {
-//			helper.notifyErrorListeners("cannot parse the boolean value: \'" +  ctx.getText() + "\'", ctx.getStart());
-//			return;
-//		}
-//		int linePosition = ctx.start.getLine();
-//		int charPosition = ctx.start.getCharPositionInLine();
-//		ctx.info.expr = new BooleanIdentifier(val, helper.getCurrentFileName(), linePosition, charPosition, linePosition, charPosition);
-//		setFileLineColumn(ctx.info.expr, ctx);
-//	}
-
 	@Override
 	public void exitConstDoubleIdExpression(ConstDoubleIdExpressionContext ctx) {
 		try {
@@ -452,7 +461,7 @@ public class DmlSyntacticValidator implements DmlListener
 			setFileLineColumn(ctx.info.expr, ctx);
 		}
 		catch(Exception e) {
-			helper.notifyErrorListeners("cannot parse the double value: \'" +  ctx.getText() + "\'", ctx.getStart());
+			helper.notifyErrorListeners("cannot parse the float value: \'" +  ctx.getText() + "\'", ctx.getStart());
 			return;
 		}
 	}
@@ -467,7 +476,7 @@ public class DmlSyntacticValidator implements DmlListener
 			setFileLineColumn(ctx.info.expr, ctx);
 		}
 		catch(Exception e) {
-			helper.notifyErrorListeners("cannot parse the integer value: \'" +  ctx.getText() + "\'", ctx.getStart());
+			helper.notifyErrorListeners("cannot parse the int value: \'" +  ctx.getText() + "\'", ctx.getStart());
 			return;
 		}
 	}
@@ -502,13 +511,6 @@ public class DmlSyntacticValidator implements DmlListener
 		int col = ctx.start.getCharPositionInLine();
 		ctx.info.expr.setAllPositions(helper.getCurrentFileName(), line, col, line, col);
 		setFileLineColumn(ctx.info.expr, ctx);
-//		if(ctx.getChild(0) instanceof DataIdentifierContext) {
-//			ctx.info.expr = ctx.dataIdentifier().dataInfo.expr;
-//		}
-//		else {
-//			String msg = "cannot evaluate data expression ... strange";
-//			helper.notifyErrorListeners(msg, ctx.start);
-//		}
 	}
 	
 	@Override
@@ -517,6 +519,13 @@ public class DmlSyntacticValidator implements DmlListener
 		// By default, it assigns to a data type
 		ctx.dataInfo.expr = new DataIdentifier(ctx.getText());
 		setFileLineColumn(ctx.dataInfo.expr, ctx);
+	}
+	
+	
+	private Expression incrementByOne(Expression expr, ParserRuleContext ctx) {
+		// For maintaining semantic consistency, we have decided to keep 1-based indexing
+		// If in future, PyDML becomes more popular than DML, this can be switched.
+		return expr;
 	}
 	
 	@Override
@@ -540,12 +549,12 @@ public class DmlSyntacticValidator implements DmlListener
 			}
 			else if(isRowLower && isRowUpper) {
 				// both set
-				rowIndices.add(ctx.rowLower.info.expr);
+				rowIndices.add(incrementByOne(ctx.rowLower.info.expr, ctx));
 				rowIndices.add(ctx.rowUpper.info.expr);
 			}
 			else if(isRowLower && !isRowUpper) {
 				// only row set
-				rowIndices.add(ctx.rowLower.info.expr);
+				rowIndices.add(incrementByOne(ctx.rowLower.info.expr, ctx));
 			}
 			else {
 				helper.notifyErrorListeners("incorrect index expression for row", ctx.start);
@@ -557,49 +566,16 @@ public class DmlSyntacticValidator implements DmlListener
 				colIndices.add(null); colIndices.add(null);
 			}
 			else if(isColLower && isColUpper) {
-				colIndices.add(ctx.colLower.info.expr);
+				colIndices.add(incrementByOne(ctx.colLower.info.expr, ctx));
 				colIndices.add(ctx.colUpper.info.expr);
 			}
 			else if(isColLower && !isColUpper) {
-				colIndices.add(ctx.colLower.info.expr);
+				colIndices.add(incrementByOne(ctx.colLower.info.expr, ctx));
 			}
 			else {
 				helper.notifyErrorListeners("incorrect index expression for column", ctx.start);
 				return;
 			}
-			
-			
-//			boolean rowIndexLowerSet = false;
-//			boolean colIndexLowerSet = false;
-//			
-//			if(ctx.rowLower != null && !ctx.rowLower.isEmpty() && (ctx.rowLower.info.expr != null)) {
-//				rowIndices.add(ctx.rowLower.info.expr);
-//				rowIndexLowerSet = true;
-//			}
-//			else {
-//				rowIndices.add(null);
-//			}
-//			if(ctx.rowUpper != null && !ctx.rowUpper.isEmpty() && (ctx.rowUpper.info.expr != null)) {
-//				rowIndices.add(ctx.rowUpper.info.expr);
-//				if(!rowIndexLowerSet) {
-//					helper.notifyErrorListeners("incorrect index expression for row", ctx.start);
-//					return;
-//				}
-//			}
-//			if(ctx.colLower != null && !ctx.colLower.isEmpty() && (ctx.colLower.info.expr != null)) {
-//				colIndices.add(ctx.colLower.info.expr);
-//				colIndexLowerSet = true;
-//			}
-//			else {
-//				colIndices.add(null);
-//			}
-//			if(ctx.colUpper != null && !ctx.colUpper.isEmpty() && (ctx.colUpper.info.expr != null)) {
-//				colIndices.add(ctx.colUpper.info.expr);
-//				if(!colIndexLowerSet) {
-//					helper.notifyErrorListeners("incorrect index expression for column", ctx.start);
-//					return;
-//				}
-//			}
 			exprList.add(rowIndices);
 			exprList.add(colIndices);
 			((IndexedIdentifier) ctx.dataInfo.expr).setIndices(exprList);
@@ -612,51 +588,48 @@ public class DmlSyntacticValidator implements DmlListener
 	
 	private ConstIdentifier getConstIdFromString(String varValue, Token start) {
 		// Both varName and varValue are correct
-				int linePosition = start.getLine();
-				int charPosition = start.getCharPositionInLine();
+		int linePosition = start.getLine();
+		int charPosition = start.getCharPositionInLine();
+		try {
+			long val = Long.parseLong(varValue);
+			return new IntIdentifier(val, helper.getCurrentFileName(), linePosition, charPosition, linePosition, charPosition);
+		}
+		catch(Exception e) {
+			try {
+				double val = Double.parseDouble(varValue);
+				return new DoubleIdentifier(val, helper.getCurrentFileName(), linePosition, charPosition, linePosition, charPosition);
+			}
+			catch(Exception e1) {
 				try {
-					long val = Long.parseLong(varValue);
-					return new IntIdentifier(val, helper.getCurrentFileName(), linePosition, charPosition, linePosition, charPosition);
-				}
-				catch(Exception e) {
-					try {
-						double val = Double.parseDouble(varValue);
-						return new DoubleIdentifier(val, helper.getCurrentFileName(), linePosition, charPosition, linePosition, charPosition);
+					if(varValue.compareTo("True") == 0 || varValue.compareTo("False") == 0) {
+						boolean val = false;
+						if(varValue.compareTo("True") == 0) {
+							val = true;
+						}
+						return new BooleanIdentifier(val, helper.getCurrentFileName(), linePosition, charPosition, linePosition, charPosition);
 					}
-					catch(Exception e1) {
-						try {
-							if(varValue.compareTo("TRUE") == 0 || varValue.compareTo("FALSE") == 0) {
-								boolean val = false;
-								if(varValue.compareTo("TRUE") == 0) {
-									val = true;
-								}
-								return new BooleanIdentifier(val, helper.getCurrentFileName(), linePosition, charPosition, linePosition, charPosition);
-							}
-							else {
-								String val = "";
-								String text = varValue;
-								if(	(text.startsWith("\"") && text.endsWith("\"")) ||
-									(text.startsWith("\'") && text.endsWith("\'"))) {
-									if(text.length() > 2) {
-										val = text.substring(1, text.length()-1);
-									}
-								}
-								else {
-									val = text;
-									// the commandline parameters can be passed without any quotes
-//									helper.notifyErrorListeners("something wrong while parsing string ... strange", start);
-//									return null;
-								}
-								return new StringIdentifier(val, helper.getCurrentFileName(), linePosition, charPosition, linePosition, charPosition);
+					else {
+						String val = "";
+						String text = varValue;
+						if(	(text.startsWith("\"") && text.endsWith("\"")) ||
+							(text.startsWith("\'") && text.endsWith("\'"))) {
+							if(text.length() > 2) {
+								val = text.substring(1, text.length()-1);
 							}
 						}
-						catch(Exception e3) {
-							helper.notifyErrorListeners("unable to cast the commandline parameter into int/double/boolean/string", start);
-							return null;
+						else {
+							// the commandline parameters can be passed without any quotes
+							val = text;
 						}
+						return new StringIdentifier(val, helper.getCurrentFileName(), linePosition, charPosition, linePosition, charPosition);
 					}
 				}
-				
+				catch(Exception e3) {
+					helper.notifyErrorListeners("unable to cast the commandline parameter into int/float/bool/str", start);
+					return null;
+				}
+			}
+		}			
 	}
 	
 	private void fillExpressionInfoCommandLineParameters(String varName, ExpressionInfo dataInfo, Token start) {
@@ -729,7 +702,7 @@ public class DmlSyntacticValidator implements DmlListener
 	// --------------------------------------------------------------------
 	
 	@Override
-	public void exitImportStatement(ImportStatementContext ctx) 
+	public void exitImportStatement(ImportStatementContext ctx)
 	{
 		//prepare import filepath
 		String filePath = ctx.filePath.getText();
@@ -747,7 +720,7 @@ public class DmlSyntacticValidator implements DmlListener
 		
 		DMLProgram prog = null;
 		try {
-			prog = (new DMLParserWrapper()).doParse(filePath, null, argVals);
+			prog = (new PyDMLParserWrapper()).doParse(filePath, null, argVals);
 		} catch (ParseException e) {
 			helper.notifyErrorListeners("Exception found during importing a program from file " + filePath, ctx.start);
 			return;
@@ -829,7 +802,6 @@ public class DmlSyntacticValidator implements DmlListener
 			int line = ctx.start.getLine();
 			int col = ctx.start.getCharPositionInLine();
 			ctx.info.stmt = new PrintStatement(functionName, expr, line, col, line, col);
-			setFileLineColumn(ctx.info.stmt, ctx);
 		} catch (LanguageException e) {
 			helper.notifyErrorListeners("cannot process " + functionName + "() function", ctx.start);
 			return;
@@ -839,61 +811,541 @@ public class DmlSyntacticValidator implements DmlListener
 	private void setOutputStatement(FunctionCallAssignmentStatementContext ctx) {
 		ArrayList<ParameterExpression> paramExpression = helper.getParameterExpressionList(ctx.paramExprs);
 		if(paramExpression.size() < 2){
-			helper.notifyErrorListeners("incorrect usage of write function (atleast 2 arguments required)", ctx.start);
+			helper.notifyErrorListeners("incorrect usage of load function (atleast 2 arguments required)", ctx.start);
 			return;
 		}
 		if(paramExpression.get(0).getExpr() instanceof DataIdentifier) {
-			//  && paramExpression.get(0).getName() == null
-			// correct usage of identifier
-			// if(paramExpression.get(1).getName() == null) {
-				String fileName = helper.getCurrentFileName();
-				int line = ctx.start.getLine();
-				int col = ctx.start.getCharPositionInLine();
-				HashMap<String, Expression> varParams = new HashMap<String, Expression>();
-				varParams.put(DataExpression.IO_FILENAME, paramExpression.get(1).getExpr());
-				for(int i = 2; i < paramExpression.size(); i++) {
-					// DataExpression.FORMAT_TYPE, DataExpression.DELIM_DELIMITER, DataExpression.DELIM_HAS_HEADER_ROW,  DataExpression.DELIM_SPARSE
-					varParams.put(paramExpression.get(i).getName(), paramExpression.get(i).getExpr());
-				}
-				
-				DataExpression  dataExpression = new DataExpression(DataOp.WRITE, varParams, fileName, line, col, line, col);
-				ctx.info.stmt = new  OutputStatement((DataIdentifier) paramExpression.get(0).getExpr(), DataOp.WRITE, fileName, line, col, line, col);
-				setFileLineColumn(ctx.info.stmt, ctx);
-				((OutputStatement)ctx.info.stmt).setExprParams(dataExpression);
-				return;
-			//}
+			String fileName = helper.getCurrentFileName();
+			int line = ctx.start.getLine();
+			int col = ctx.start.getCharPositionInLine();
+			HashMap<String, Expression> varParams = new HashMap<String, Expression>();
+			varParams.put(DataExpression.IO_FILENAME, paramExpression.get(1).getExpr());
+			for(int i = 2; i < paramExpression.size(); i++) {
+				// DataExpression.FORMAT_TYPE, DataExpression.DELIM_DELIMITER, DataExpression.DELIM_HAS_HEADER_ROW,  DataExpression.DELIM_SPARSE
+				varParams.put(paramExpression.get(i).getName(), paramExpression.get(i).getExpr());
+			}
+			
+			DataExpression  dataExpression = new DataExpression(DataOp.WRITE, varParams, fileName, line, col, line, col);
+			ctx.info.stmt = new  OutputStatement((DataIdentifier) paramExpression.get(0).getExpr(), DataOp.WRITE, fileName, line, col, line, col);
+			setFileLineColumn(ctx.info.stmt, ctx);
+			((OutputStatement)ctx.info.stmt).setExprParams(dataExpression);
+			return;
 		}
 		
 		helper.notifyErrorListeners("incorrect usage of write function", ctx.start);
 		return;
 		
 	}
+
+	private boolean inDefaultNamespace(String namespace) {
+		return namespace.compareTo(DMLProgram.DEFAULT_NAMESPACE) == 0;
+	}
+	
+	// Returns 0, 1 or -1 (for error)
+	private int getAxis(ParameterizedExpressionContext ctx) {
+		if(ctx.paramName != null && ctx.paramName.getText() != null && !ctx.paramName.getText().isEmpty()) {
+			if(ctx.paramName.getText().compareTo("axis") != 0) {
+				return -1;
+			}
+		}
+		
+		String val = ctx.paramVal.getText();
+		if(val != null && val.compareTo("0") == 0) {
+			return 0;
+		}
+		else if(val != null && val.compareTo("1") == 0) {
+			return 1;
+		}
+		
+		return -1;
+	}
+	
+	private String getPythonAggFunctionNames(String functionName, int axis) {
+		if(axis != 0 && axis != 1) {
+			return functionName;
+		}
+		// axis=0 maps to column-wise computation and axis=1 maps to row-wise computation
+		
+		if(functionName.compareTo("sum") == 0) {
+			return axis == 0 ? "colSums" : "rowSums"; 
+		}
+		else if(functionName.compareTo("mean") == 0) {
+			return axis == 0 ? "colMeans" : "rowMeans"; 
+		}
+		else if(functionName.compareTo("avg") == 0) {
+			return axis == 0 ? "colMeans" : "rowMeans";
+		}
+		else if(functionName.compareTo("max") == 0) {
+			return axis == 0 ? "colMaxs" : "rowMaxs";
+		}
+		else if(functionName.compareTo("min") == 0) {
+			return axis == 0 ? "colMins" : "rowMins";
+		}
+		else if(functionName.compareTo("argmin") == 0) {
+			return axis == 0 ? "Not Supported" : "rowIndexMin";
+		}
+		else if(functionName.compareTo("argmax") == 0) {
+			return axis == 0 ? "Not Supported" : "rowIndexMax";
+		}
+		else if(functionName.compareTo("cumsum") == 0) {
+			return axis == 0 ?  "cumsum" : "Not Supported";
+		}
+		else if(functionName.compareTo("transpose") == 0) {
+			return axis == 0 ?  "Not Supported" : "Not Supported";
+		}
+		else if(functionName.compareTo("trace") == 0) {
+			return axis == 0 ?  "Not Supported" : "Not Supported";
+		}
+		else {
+			return functionName;
+		}
+	}
+	
+	private ConvertedDMLSyntax convertPythonBuiltinFunctionToDMLSyntax(String namespace, String functionName, ArrayList<ParameterExpression> paramExpression, 
+		List<ParameterizedExpressionContext> paramCtx, Token fnName, String fileName, int line, int col) {
+		// ===========================================================================================
+		// Check function name, namespace, parameters (#params & possible values) and throw useful hints
+		if(inDefaultNamespace(namespace) && functionName.compareTo("len") == 0) {
+			if(paramExpression.size() != 1) {
+				helper.notifyErrorListeners("The builtin function \'" + functionName + "\' accepts 1 arguments", fnName);
+				return null;
+			}
+			functionName = "length";
+		}
+		else if(functionName.compareTo("sum") == 0 || functionName.compareTo("mean") == 0 || functionName.compareTo("avg") == 0 ||
+				functionName.compareTo("min") == 0 || functionName.compareTo("max") == 0  || 
+				functionName.compareTo("argmax") == 0 || functionName.compareTo("argmin") == 0 ||
+				functionName.compareTo("cumsum") == 0 || functionName.compareTo("transpose") == 0 || functionName.compareTo("trace") == 0) {
+			// 0 maps row-wise computation and 1 maps to column-wise computation
+			
+			// can mean sum of all cells or row-wise or columnwise sum
+			if(inDefaultNamespace(namespace) && paramExpression.size() == 1) {
+				// sum(x) => sum(x)
+				// otherwise same function name
+				if(functionName.compareTo("avg") == 0) {
+					functionName = "mean";
+				}
+				else if(functionName.compareTo("transpose") == 0) {
+					functionName = "t";
+				}
+				else if(functionName.compareTo("argmax") == 0 || functionName.compareTo("argmin") == 0 || functionName.compareTo("cumsum") == 0) {
+					helper.notifyErrorListeners("The builtin function \'" + functionName + "\' for entire matrix is not supported", fnName);
+					return null;
+				}
+			}
+			else if(!inDefaultNamespace(namespace) && paramExpression.size() == 0) {
+				// x.sum() => sum(x)
+				paramExpression = new ArrayList<ParameterExpression>();
+				paramExpression.add(new ParameterExpression(null, new DataIdentifier(namespace)));
+				// otherwise same function name
+				if(functionName.compareTo("avg") == 0) {
+					functionName = "mean";
+				}
+				else if(functionName.compareTo("transpose") == 0) {
+					functionName = "t";
+				}
+				else if(functionName.compareTo("argmax") == 0 || functionName.compareTo("argmin") == 0 || functionName.compareTo("cumsum") == 0) {
+					helper.notifyErrorListeners("The builtin function \'" + functionName + "\' for entire matrix is not supported", fnName);
+					return null;
+				}
+			}
+			else if(inDefaultNamespace(namespace) && paramExpression.size() == 2) {
+				// sum(x, axis=1) => rowSums(x)
+				int axis = getAxis(paramCtx.get(1));
+				if(axis == -1 && (functionName.compareTo("min") == 0 || functionName.compareTo("max") == 0 )) {
+					// Do nothing
+					// min(2, 3)
+				}
+				else if(axis == -1) {
+					helper.notifyErrorListeners("The builtin function \'" + functionName + "\' for given arguments is not supported", fnName);
+					return null;
+				}
+				else {
+					ArrayList<ParameterExpression> temp = new ArrayList<ParameterExpression>();
+					temp.add(paramExpression.get(0));
+					paramExpression = temp;
+					functionName = getPythonAggFunctionNames(functionName, axis);
+					if(functionName.compareTo("Not Supported") == 0) {
+						helper.notifyErrorListeners("The builtin function \'" + functionName + "\' for given arguments is not supported", fnName);
+						return null;
+					}
+				}
+			}
+			else if(!inDefaultNamespace(namespace) && paramExpression.size() == 1) {
+				// x.sum(axis=1) => rowSums(x)
+				int axis = getAxis(paramCtx.get(0));
+				 if(axis == -1) {
+					 helper.notifyErrorListeners("The builtin function \'" + functionName + "\' for given arguments is not supported", fnName);
+					 return null;
+				 }
+				 else {
+					 paramExpression = new ArrayList<ParameterExpression>();
+					 paramExpression.add(new ParameterExpression(null, new DataIdentifier(namespace)));
+					 functionName = getPythonAggFunctionNames(functionName, axis);
+					 if(functionName.compareTo("Not Supported") == 0) {
+						 helper.notifyErrorListeners("The builtin function \'" + functionName + "\' for given arguments is not supported", fnName);
+						 return null;
+					 }
+				 }
+			}
+			else {
+				helper.notifyErrorListeners("Incorrect number of arguments for the builtin function \'" + functionName + "\'.", fnName);
+				return null;
+			}
+			namespace = DMLProgram.DEFAULT_NAMESPACE;
+		}
+		else if(inDefaultNamespace(namespace) && functionName.compareTo("concatenate") == 0) {
+			if(paramExpression.size() != 2) {
+				helper.notifyErrorListeners("The builtin function \'" + functionName + "\' accepts 2 arguments (Note: concatenate append columns of two matrices)", fnName);
+				return null;
+			}
+			functionName = "append";
+			namespace = DMLProgram.DEFAULT_NAMESPACE;
+		}
+		else if(inDefaultNamespace(namespace) && functionName.compareTo("minimum") == 0) {
+			if(paramExpression.size() != 2) {
+				helper.notifyErrorListeners("The builtin function \'" + functionName + "\' accepts 2 arguments", fnName);
+				return null;
+			}
+			functionName = "min";
+			namespace = DMLProgram.DEFAULT_NAMESPACE;
+		}
+		else if(inDefaultNamespace(namespace) && functionName.compareTo("maximum") == 0) {
+			if(paramExpression.size() != 2) {
+				helper.notifyErrorListeners("The builtin function \'" + functionName + "\' accepts 2 arguments", fnName);
+				return null;
+			}
+			functionName = "max";
+			namespace = DMLProgram.DEFAULT_NAMESPACE;
+		}
+		else if(!inDefaultNamespace(namespace) && functionName.compareTo("shape") == 0) {
+			if(paramExpression.size() != 1) {
+				helper.notifyErrorListeners("The builtin function \'" + functionName + "\' accepts only 1 argument (0 or 1)", fnName);
+				return null;
+			}
+			
+			int axis = getAxis(paramCtx.get(0));
+			if(axis == -1) {
+				helper.notifyErrorListeners("The builtin function \'" + functionName + "\' accepts only 1 argument (0 or 1)", fnName);
+				return null; 
+			}
+			paramExpression = new ArrayList<ParameterExpression>();
+			paramExpression.add(new ParameterExpression(null, new DataIdentifier(namespace)));
+			namespace = DMLProgram.DEFAULT_NAMESPACE;
+			if(axis == 0) {
+				functionName = "nrow";
+			}
+			else if(axis == 1) {
+				functionName = "ncol";
+			}
+		}
+		else if(inDefaultNamespace(namespace) && functionName.compareTo("random.normal") == 0) {
+			if(paramExpression.size() != 3) {
+				helper.notifyErrorListeners("The builtin function \'" + functionName + "\' accepts exactly 3 arguments (number of rows, number of columns, sparsity)", fnName);
+				return null;
+			}
+			paramExpression.get(0).setName("rows");
+			paramExpression.get(1).setName("cols");
+			paramExpression.get(2).setName("sparsity");
+			paramExpression.add(new org.apache.sysml.parser.ParameterExpression("pdf", new StringIdentifier("normal", fileName, line, col, line, col)));
+			functionName = "rand";
+			namespace = DMLProgram.DEFAULT_NAMESPACE;
+		}
+		else if(inDefaultNamespace(namespace) && functionName.compareTo("random.uniform") == 0) {
+			if(paramExpression.size() != 5) {
+				helper.notifyErrorListeners("The builtin function \'" + functionName + "\' accepts exactly 5 arguments (number of rows, number of columns, sparsity, min, max)", fnName);
+				return null;
+			}
+			paramExpression.get(0).setName("rows");
+			paramExpression.get(1).setName("cols");
+			paramExpression.get(2).setName("sparsity");
+			paramExpression.get(3).setName("min");
+			paramExpression.get(4).setName("max");
+			paramExpression.add(new org.apache.sysml.parser.ParameterExpression("pdf", new StringIdentifier("uniform", fileName, line, col, line, col)));
+			functionName = "rand";
+			namespace = DMLProgram.DEFAULT_NAMESPACE;
+		}
+		else if(inDefaultNamespace(namespace) && functionName.compareTo("full") == 0) {
+			if(paramExpression.size() != 3) {
+				helper.notifyErrorListeners("The builtin function \'" + functionName + "\' accepts exactly 3 arguments (constant float value, number of rows, number of columns)", fnName);
+				return null;
+			}
+			paramExpression.get(1).setName("rows");
+			paramExpression.get(2).setName("cols");
+			functionName = "matrix";
+			namespace = DMLProgram.DEFAULT_NAMESPACE;
+		}
+		else if(inDefaultNamespace(namespace) && functionName.compareTo("matrix") == 0) {
+			// This can either be string initializer or as.matrix function
+			if(paramExpression.size() != 1) {
+				helper.notifyErrorListeners("The builtin function \'" + functionName + "\' accepts exactly 1 argument (either str or float value)", fnName);
+				return null;
+			}
+			
+			if(paramExpression.get(0).getExpr() instanceof StringIdentifier) {
+				String initializerString = ((StringIdentifier)paramExpression.get(0).getExpr()).getValue().trim();
+				if(!initializerString.startsWith("[") || !initializerString.endsWith("]")) {
+					helper.notifyErrorListeners("Incorrect initializer string for builtin function \'" + functionName + "\' (Eg: matrix(\"[1 2 3; 4 5 6]\"))", fnName);
+					return null;
+				}
+				int rows = StringUtils.countMatches(initializerString, ";") + 1;
+				
+				// Make sure user doesnot have pretty string
+				initializerString = initializerString.replaceAll("; ", ";");
+				initializerString = initializerString.replaceAll(" ;", ";");
+				initializerString = initializerString.replaceAll("\\[ ", "\\[");
+				initializerString = initializerString.replaceAll(" \\]", "\\]");
+				
+				// Each row has ncol-1 spaces
+				// #spaces = nrow * (ncol-1)
+				// ncol = (#spaces / nrow) + 1 
+				int cols = (StringUtils.countMatches(initializerString, " ") / rows) + 1;
+				
+				initializerString = initializerString.replaceAll(";", " ");
+				initializerString = initializerString.replaceAll("\\[", "");
+				initializerString = initializerString.replaceAll("\\]", "");
+				paramExpression = new ArrayList<ParameterExpression>();
+				paramExpression.add(new ParameterExpression(null, new StringIdentifier(initializerString, fileName, line, col, line, col)));
+				paramExpression.add(new ParameterExpression("rows", new IntIdentifier(rows, fileName, line, col, line, col)));
+				paramExpression.add(new ParameterExpression("cols", new IntIdentifier(cols, fileName, line, col, line, col)));
+			}
+			else {
+				functionName = "as.matrix";
+			}
+			namespace = DMLProgram.DEFAULT_NAMESPACE;
+		}
+		else if(inDefaultNamespace(namespace) && functionName.compareTo("scalar") == 0) {
+			if(paramExpression.size() != 1) {
+				helper.notifyErrorListeners("The builtin function \'" + functionName + "\' accepts exactly 1 argument", fnName);
+				return null;
+			}
+			functionName = "as.scalar";
+			namespace = DMLProgram.DEFAULT_NAMESPACE;
+		}
+		else if(inDefaultNamespace(namespace) && functionName.compareTo("float") == 0) {
+			if(paramExpression.size() != 1) {
+				helper.notifyErrorListeners("The builtin function \'" + functionName + "\' accepts exactly 1 argument", fnName);
+				return null;
+			}
+			functionName = "as.double";
+			namespace = DMLProgram.DEFAULT_NAMESPACE;
+		}
+		else if(inDefaultNamespace(namespace) && functionName.compareTo("int") == 0) {
+			if(paramExpression.size() != 1) {
+				helper.notifyErrorListeners("The builtin function \'" + functionName + "\' accepts exactly 1 argument", fnName);
+				return null;
+			}
+			functionName = "as.integer";
+			namespace = DMLProgram.DEFAULT_NAMESPACE;
+		}
+		else if(inDefaultNamespace(namespace) && functionName.compareTo("bool") == 0) {
+			if(paramExpression.size() != 1) {
+				helper.notifyErrorListeners("The builtin function \'" + functionName + "\' accepts exactly 1 argument", fnName);
+				return null;
+			}
+			functionName = "as.logical";
+			namespace = DMLProgram.DEFAULT_NAMESPACE;
+		}
+		else if(!inDefaultNamespace(namespace) && functionName.compareTo("reshape") == 0) {
+			if(paramExpression.size() != 2) {
+				helper.notifyErrorListeners("The builtin function \'" + functionName + "\' accepts exactly 2 arguments (number of rows, number of columns)", fnName);
+				return null;
+			}
+			paramExpression.get(0).setName("rows");
+			paramExpression.get(1).setName("cols");
+			
+			ArrayList<ParameterExpression> temp = new ArrayList<ParameterExpression>();
+			temp.add(new ParameterExpression(null, new DataIdentifier(namespace)));
+			temp.add(paramExpression.get(0));
+			temp.add(paramExpression.get(1));
+			paramExpression = temp;
+			
+			functionName = "matrix";
+			namespace = DMLProgram.DEFAULT_NAMESPACE;
+		}
+		else if(inDefaultNamespace(namespace) && functionName.compareTo("removeEmpty") == 0) {
+			if(paramExpression.size() != 2) {
+				helper.notifyErrorListeners("The builtin function \'" + functionName + "\' accepts exactly 2 arguments (matrix, axis=0 or 1)", fnName);
+				return null;
+			}
+			int axis = getAxis(paramCtx.get(1));
+			if(axis == -1) {
+				helper.notifyErrorListeners("The builtin function \'" + functionName + "\' accepts exactly 2 arguments (matrix, axis=0 or 1)", fnName);
+				return null; 
+			}
+			StringIdentifier marginVal = null;
+			if(axis == 0) {
+				marginVal = new StringIdentifier("rows", fileName, line, col, line, col);
+			}
+			else {
+				marginVal = new StringIdentifier("cols", fileName, line, col, line, col);
+			}
+			paramExpression.get(0).setName("target");
+			paramExpression.get(1).setName("margin");
+			paramExpression.get(1).setExpr(marginVal);
+			functionName = "removeEmpty";
+			namespace = DMLProgram.DEFAULT_NAMESPACE;
+		}
+		else if(inDefaultNamespace(namespace) && functionName.compareTo("replace") == 0) {
+			if(paramExpression.size() != 3) {
+				helper.notifyErrorListeners("The builtin function \'" + functionName + "\' accepts exactly 3 arguments (matrix, scalar value that should be replaced (pattern), scalar value (replacement))", fnName);
+				return null;
+			}
+			paramExpression.get(0).setName("target");
+			paramExpression.get(1).setName("pattern");
+			paramExpression.get(2).setName("replacement");
+			functionName = "replace";
+			namespace = DMLProgram.DEFAULT_NAMESPACE;
+		}
+		else if(inDefaultNamespace(namespace) && functionName.compareTo("range") == 0) {
+			if(paramExpression.size() != 3) {
+				helper.notifyErrorListeners("The builtin function \'" + functionName + "\' accepts exactly 3 arguments (matrix, scalar value that should be replaced (pattern), scalar value (replacement))", fnName);
+				return null;
+			}
+			functionName = "seq";
+			namespace = DMLProgram.DEFAULT_NAMESPACE;
+		}
+		else if(inDefaultNamespace(namespace) && functionName.compareTo("norm.cdf") == 0) {
+			if(paramExpression.size() != 3) {
+				helper.notifyErrorListeners("The builtin function \'" + functionName + "\' accepts exactly 3 arguments (target, mean, sd)", fnName);
+				return null;
+			}
+			functionName = "cumulativeProbability";
+			paramExpression.get(0).setName("target");
+			paramExpression.get(1).setName("mean");
+			paramExpression.get(2).setName("sd");
+			paramExpression.add(new ParameterExpression("dist", new StringIdentifier("normal", fileName, line, col, line, col)));
+			namespace = DMLProgram.DEFAULT_NAMESPACE;
+		}
+		else if(inDefaultNamespace(namespace) && functionName.compareTo("expon.cdf") == 0) {
+			if(paramExpression.size() != 2) {
+				helper.notifyErrorListeners("The builtin function \'" + functionName + "\' accepts exactly 2 arguments (target, mean)", fnName);
+				return null;
+			}
+			functionName = "cumulativeProbability";
+			paramExpression.get(0).setName("target");
+			paramExpression.get(1).setName("mean");
+			paramExpression.add(new ParameterExpression("dist", new StringIdentifier("exp", fileName, line, col, line, col)));
+			namespace = DMLProgram.DEFAULT_NAMESPACE;
+		}
+		else if(inDefaultNamespace(namespace) && functionName.compareTo("chi.cdf") == 0) {
+			if(paramExpression.size() != 2) {
+				helper.notifyErrorListeners("The builtin function \'" + functionName + "\' accepts exactly 2 arguments (target, df)", fnName);
+				return null;
+			}
+			functionName = "cumulativeProbability";
+			paramExpression.get(0).setName("target");
+			paramExpression.get(1).setName("df");
+			paramExpression.add(new ParameterExpression("dist", new StringIdentifier("chisq", fileName, line, col, line, col)));
+			namespace = DMLProgram.DEFAULT_NAMESPACE;
+		}
+		else if(inDefaultNamespace(namespace) && functionName.compareTo("f.cdf") == 0) {
+			if(paramExpression.size() != 3) {
+				helper.notifyErrorListeners("The builtin function \'" + functionName + "\' accepts exactly 3 arguments (target, df1, df2)", fnName);
+				return null;
+			}
+			functionName = "cumulativeProbability";
+			paramExpression.get(0).setName("target");
+			paramExpression.get(1).setName("df1");
+			paramExpression.get(2).setName("df2");
+			paramExpression.add(new ParameterExpression("dist", new StringIdentifier("f", fileName, line, col, line, col)));
+			namespace = DMLProgram.DEFAULT_NAMESPACE;
+		}
+		else if(inDefaultNamespace(namespace) && functionName.compareTo("t.cdf") == 0) {
+			if(paramExpression.size() != 2) {
+				helper.notifyErrorListeners("The builtin function \'" + functionName + "\' accepts exactly 2 arguments (target, df)", fnName);
+				return null;
+			}
+			functionName = "cumulativeProbability";
+			paramExpression.get(0).setName("target");
+			paramExpression.get(1).setName("df");
+			paramExpression.add(new ParameterExpression("dist", new StringIdentifier("t", fileName, line, col, line, col)));
+			namespace = DMLProgram.DEFAULT_NAMESPACE;
+		}
+		else if(inDefaultNamespace(namespace) && functionName.compareTo("percentile") == 0) {
+			if(paramExpression.size() != 2 && paramExpression.size() != 3) {
+				helper.notifyErrorListeners("The builtin function \'" + functionName + "\' accepts either 2 or 3 arguments", fnName);
+				return null;
+			}
+			functionName = "quantile";
+			namespace = DMLProgram.DEFAULT_NAMESPACE;
+		}
+		else if(inDefaultNamespace(namespace) && functionName.compareTo("arcsin") == 0) {
+			functionName = "asin";
+		}
+		else if(inDefaultNamespace(namespace) && functionName.compareTo("arccos") == 0) {
+			functionName = "acos";
+		}
+		else if(inDefaultNamespace(namespace) && functionName.compareTo("arctan") == 0) {
+			functionName = "atan";
+		}
+		else if(inDefaultNamespace(namespace) && functionName.compareTo("load") == 0) {
+			functionName = "read";
+		}
+		else if(inDefaultNamespace(namespace) && functionName.compareTo("eigen") == 0) {
+			functionName = "eig";
+		}
+		else if(inDefaultNamespace(namespace) && functionName.compareTo("power") == 0) {
+			if(paramExpression.size() != 2) {
+				helper.notifyErrorListeners("The builtin function \'" + functionName + "\' accepts exactly 2 arguments", fnName);
+				return null;
+			}
+		}
+		else if(inDefaultNamespace(namespace) && functionName.compareTo("dot") == 0) {
+			if(paramExpression.size() != 2) {
+				helper.notifyErrorListeners("The builtin function \'" + functionName + "\' accepts exactly 2 arguments", fnName);
+				return null;
+			}
+		}
+		
+		ConvertedDMLSyntax retVal = new ConvertedDMLSyntax();
+		retVal.namespace = namespace;
+		retVal.functionName = functionName;
+		retVal.paramExpression = paramExpression;
+		return retVal;
+	}
+
+	class ConvertedDMLSyntax {
+		public String namespace;
+		public String functionName;
+		public ArrayList<ParameterExpression> paramExpression;
+	};
+	
+	private Expression getOperatorExpression(String namespace, String functionName, ArrayList<ParameterExpression> paramExpression) {
+		String dmlOperator = null;
+		
+		if(inDefaultNamespace(namespace) && functionName.compareTo("dot") == 0) {
+			if(paramExpression.size() == 2) {
+				dmlOperator = "%*%";
+			}
+		}
+
+		if(dmlOperator != null) {
+			Expression.BinaryOp bop = Expression.getBinaryOp(dmlOperator);
+			Expression expr = new BinaryExpression(bop);
+			((BinaryExpression)expr).setLeft(paramExpression.get(0).getExpr());
+			((BinaryExpression)expr).setRight(paramExpression.get(1).getExpr());
+			return expr;
+		}
+		
+		return null;
+	}
 	
 	@Override
 	public void exitFunctionCallAssignmentStatement(FunctionCallAssignmentStatementContext ctx) {
-		String fullyQualifiedFunctionName = ctx.name.getText();
-		String [] fnNames = fullyQualifiedFunctionName.split("::");
-		String functionName = "";
-		String namespace = "";
-		if(fnNames.length == 1) {
-			namespace = DMLProgram.DEFAULT_NAMESPACE;
-			functionName = fnNames[0].trim();
-		}
-		else if(fnNames.length == 2) {
-			namespace = fnNames[0].trim();
-			functionName = fnNames[1].trim();
-		}
-		else {
-			helper.notifyErrorListeners("incorrect function name", ctx.name);
+		ArrayList<String> names = helper.getQualifiedNames(ctx.name.getText());
+		if(names == null) {
+			helper.notifyErrorListeners("incorrect function name (only namespace.functionName allowed. Hint: If you are trying to use builtin functions, you can skip the namespace)", ctx.name);
 			return;
 		}
+		String namespace = names.get(0);
+		String functionName = names.get(1);
 		
 		if((functionName.compareTo("print") == 0 || functionName.compareTo("stop") == 0 ) && namespace.compareTo(DMLProgram.DEFAULT_NAMESPACE) == 0) {
 			setPrintStatement(ctx, functionName);
 			return;
 		}
-		else if(functionName.compareTo("write") == 0
-				&& namespace.compareTo(DMLProgram.DEFAULT_NAMESPACE) == 0) {
+		else if(functionName.compareTo("save") == 0 && namespace.compareTo(DMLProgram.DEFAULT_NAMESPACE) == 0) {
 			setOutputStatement(ctx);
 			return;
 		}
@@ -908,10 +1360,18 @@ public class DmlSyntacticValidator implements DmlListener
 		int col = ctx.start.getCharPositionInLine();
 		
 		ArrayList<ParameterExpression> paramExpression = helper.getParameterExpressionList(ctx.paramExprs);
-//		if(functionName.compareTo("read") == 0 && paramExpression.size() > 0 && paramExpression.get(0).getName() == null) {
-//			paramExpression.get(0).setName(DataExpression.IO_FILENAME);
-//		}
 		
+	 	ConvertedDMLSyntax convertedSyntax = convertPythonBuiltinFunctionToDMLSyntax(namespace, functionName, paramExpression, ctx.paramExprs, ctx.name, fileName, line, col);
+		if(convertedSyntax == null) {
+			return;
+		}
+		else {
+			namespace = convertedSyntax.namespace;
+			functionName = convertedSyntax.functionName;
+			paramExpression = convertedSyntax.paramExpression;
+		}
+		
+		// ===========================================================================================
 		FunctionCallIdentifier functCall = new FunctionCallIdentifier(paramExpression);
 		try {
 			functCall.setFunctionName(functionName);
@@ -933,34 +1393,50 @@ public class DmlSyntacticValidator implements DmlListener
 			}
 		}
 		
-		if(!functionName.contains("::") || functionName.startsWith(DMLProgram.DEFAULT_NAMESPACE)) {
+		Expression operatorExpr = getOperatorExpression(namespace, functionName, paramExpression);
+		if(operatorExpr != null) {
+			setFileLineColumn(operatorExpr, ctx);
+			setAssignmentStatement(target, operatorExpr, ctx);
+			return;
+		}
+		
+		//Note: In contrast to the dml parser, namespace and function names are separated by '.' not '::'.  
+		//Hence, we have to include a whitelist of function names to handle builtins like 'as.scalar'.		
+		String[] whitelist = new String[]{"as.matrix","as.scalar","as.double","as.integer","as.logical"};
+		boolean isWhitelisted = Arrays.asList(whitelist).contains(functionName);
+		
+		if(    !functionName.contains(".") || isWhitelisted 
+			|| functionName.startsWith(DMLProgram.DEFAULT_NAMESPACE) ) 
+		{
 			// In global namespace, so it can be a builtin function
 			if(!helper.validateBuiltinFunctions(ctx)) {
 				return; // it is a built-in function and validation failed, so donot proceed ahead.
 			}
 			// Double verification: verify passed function name is a (non-parameterized) built-in function.
-			try {
+			try 
+			{
+				// builtin functions
 				BuiltinFunctionExpression bife = BuiltinFunctionExpression.getBuiltinFunctionExpression(functionName, functCall.getParamExprs(), fileName, line, col, line, col);
-				if (bife != null){
-					// It is a builtin function
+				if (bife != null) {
 					setAssignmentStatement(target, bife, ctx);
 					return;
 				}
 				
+				// parameterized builtin functions
 				ParameterizedBuiltinFunctionExpression pbife = ParameterizedBuiltinFunctionExpression.getParamBuiltinFunctionExpression(functionName, functCall.getParamExprs(), fileName, line, col, line, col);
-				if (pbife != null){
-					// It is a parameterized builtin function
+				if (pbife != null) {
 					setAssignmentStatement(target, pbife, ctx);
 					return;
 				}
 				
-				// built-in read, rand ...
+				// built-in data expressions, e.g. read
 				DataExpression dbife = DataExpression.getDataExpression(functionName, functCall.getParamExprs(), fileName, line, col, line, col);
 				if (dbife != null){
 					setAssignmentStatement(target, dbife, ctx);
 					return;
 				}
-			} catch(Exception e) {
+			} 
+			catch(Exception e) {
 				helper.notifyErrorListeners("unable to process builtin function expression " + functionName  + ":" + e.getMessage(), ctx.start);
 				return ;
 			}
@@ -969,27 +1445,46 @@ public class DmlSyntacticValidator implements DmlListener
 		setAssignmentStatement(target, functCall, ctx);
 	}
 	
+	
 	@Override
 	public void exitBuiltinFunctionExpression(BuiltinFunctionExpressionContext ctx) {
-//		if(!helper.validateBuiltinFunctions(ctx)) {
-//			return; // it is a built-in function and validation failed, so donot proceed ahead.
-//		}
 		// Double verification: verify passed function name is a (non-parameterized) built-in function.
-		String functionName = ctx.name.getText();
+		ArrayList<String> names = helper.getQualifiedNames(ctx.name.getText());
+		if(names == null) {
+			helper.notifyErrorListeners("incorrect function name (only namespace.functionName allowed. Hint: If you are trying to use builtin functions, you can skip the namespace)", ctx.name);
+			return;
+		}
+		String namespace = names.get(0);
+		String functionName = names.get(1);
+		
 		String fileName = helper.getCurrentFileName();
 		int line = ctx.start.getLine();
 		int col = ctx.start.getCharPositionInLine();
 		ArrayList<ParameterExpression> paramExpression = helper.getParameterExpressionList(ctx.paramExprs);
-//		if(functionName.compareTo("read") == 0 && paramExpression.size() > 0 && paramExpression.get(0).getName() == null) {
-//			paramExpression.get(0).setName(DataExpression.IO_FILENAME);
-//		}
+
+		ConvertedDMLSyntax convertedSyntax = convertPythonBuiltinFunctionToDMLSyntax(namespace, functionName, paramExpression, ctx.paramExprs, ctx.name, fileName, line, col);
+		if(convertedSyntax == null) {
+			return;
+		}
+		else {
+			namespace = convertedSyntax.namespace;
+			functionName = convertedSyntax.functionName;
+			paramExpression = convertedSyntax.paramExpression;
+			// System.out.println(ctx.name.getText() + ">>" + namespace + " " + functionName);
+		}
+		
+		Expression operatorExpr = getOperatorExpression(namespace, functionName, paramExpression);
+		if(operatorExpr != null) {
+			ctx.info.expr = operatorExpr;
+			setFileLineColumn(operatorExpr, ctx);
+			return;
+		}
 		
 		try {
 			BuiltinFunctionExpression bife = BuiltinFunctionExpression.getBuiltinFunctionExpression(functionName, paramExpression, fileName, line, col, line, col);
 			if (bife != null){
 				// It is a builtin function
 				ctx.info.expr = bife;
-				setFileLineColumn(ctx.info.expr, ctx);
 				return;
 			}
 			
@@ -997,7 +1492,6 @@ public class DmlSyntacticValidator implements DmlListener
 			if (pbife != null){
 				// It is a parameterized builtin function
 				ctx.info.expr = pbife;
-				setFileLineColumn(ctx.info.expr, ctx);
 				return;
 			}
 			
@@ -1005,10 +1499,8 @@ public class DmlSyntacticValidator implements DmlListener
 			DataExpression dbife = DataExpression.getDataExpression(functionName, paramExpression, fileName, line, col, line, col);
 			if (dbife != null){
 				ctx.info.expr = dbife;
-				setFileLineColumn(ctx.info.expr, ctx);
 				return;
 			}
-			
 		} catch(Exception e) {
 			helper.notifyErrorListeners("unable to process builtin function expression " + functionName + ":" + e.getMessage(), ctx.start);
 			return ;
@@ -1025,31 +1517,30 @@ public class DmlSyntacticValidator implements DmlListener
 	@Override
 	public void exitFunctionCallMultiAssignmentStatement(
 			FunctionCallMultiAssignmentStatementContext ctx) {
-		String fullyQualifiedFunctionName = ctx.name.getText();
-		String [] fnNames = fullyQualifiedFunctionName.split("::");
-		String functionName = "";
-		String namespace = "";
-		if(fnNames.length == 1) {
-			namespace = DMLProgram.DEFAULT_NAMESPACE;
-			functionName = fnNames[0].trim();
-		}
-		else if(fnNames.length == 2) {
-			namespace = fnNames[0].trim();
-			functionName = fnNames[1].trim();
-		}
-		else {
-			helper.notifyErrorListeners("incorrect function name", ctx.name);
+		ArrayList<String> names = helper.getQualifiedNames(ctx.name.getText());
+		if(names == null) {
+			helper.notifyErrorListeners("incorrect function name (only namespace.functionName allowed. Hint: If you are trying to use builtin functions, you can skip the namespace)", ctx.name);
 			return;
 		}
+		String namespace = names.get(0);
+		String functionName = names.get(1);
 		
 		String fileName = helper.getCurrentFileName();
 		int line = ctx.start.getLine();
 		int col = ctx.start.getCharPositionInLine();
 		
 		ArrayList<ParameterExpression> paramExpression = helper.getParameterExpressionList(ctx.paramExprs);
-//		if(functionName.compareTo("read") == 0 && paramExpression.size() > 0 && paramExpression.get(0).getName() == null) {
-//			paramExpression.get(0).setName(DataExpression.IO_FILENAME);
-//		}
+		ConvertedDMLSyntax convertedSyntax = convertPythonBuiltinFunctionToDMLSyntax(namespace, functionName, paramExpression, ctx.paramExprs, ctx.name, fileName, line, col);
+		if(convertedSyntax == null) {
+			return;
+		}
+		else {
+			namespace = convertedSyntax.namespace;
+			functionName = convertedSyntax.functionName;
+			paramExpression = convertedSyntax.paramExpression;
+		}
+		
+		// No need to support dot() function since it will never return multi-assignment function
 		
 		FunctionCallIdentifier functCall = new FunctionCallIdentifier(paramExpression);
 		try {
@@ -1072,11 +1563,8 @@ public class DmlSyntacticValidator implements DmlListener
 			}
 		}
 		
-		if(!functionName.contains("::") || functionName.startsWith(DMLProgram.DEFAULT_NAMESPACE)) {
+		if(!functionName.contains(".") || functionName.startsWith(DMLProgram.DEFAULT_NAMESPACE)) {
 			// In global namespace, so it can be a builtin function
-//			if(!helper.validateBuiltinFunctions(ctx)) {
-//				return; // it is a built-in function and validation failed, so donot proceed ahead.
-//			}
 			// Double verification: verify passed function name is a (non-parameterized) built-in function.
 			try {
 				BuiltinFunctionExpression bife = BuiltinFunctionExpression.getBuiltinFunctionExpression(functionName, functCall.getParamExprs(), fileName, line, col, line, col);
@@ -1109,7 +1597,7 @@ public class DmlSyntacticValidator implements DmlListener
 	}
 	
 	private StatementBlock getStatementBlock(Statement current) {
-		return DMLParserWrapper.getStatementBlock(current);
+		return PyDMLParserWrapper.getStatementBlock(current);
 	}
 	
 	@Override
@@ -1220,13 +1708,9 @@ public class DmlSyntacticValidator implements DmlListener
 	}
 	
 	
-	@Override
-	public void exitDmlprogram(DmlprogramContext ctx) { }
 	
 
 	// ----------------------------------------------------------------------
-	@Override
-	public void exitValueType(ValueTypeContext ctx) { }
 	@Override
 	public void exitMl_type(Ml_typeContext ctx) { }
 	
@@ -1245,11 +1729,11 @@ public class DmlSyntacticValidator implements DmlListener
 				dataType = paramCtx.paramType.dataType().getText();
 			}
 			
-			if(dataType.compareTo("matrix") == 0 || dataType.compareTo("Matrix") == 0) {
+			if(dataType.compareTo("matrix") == 0) {
 				// matrix
 				dataId.setDataType(DataType.MATRIX);
 			}
-			else if(dataType.compareTo("scalar") == 0 || dataType.compareTo("Scalar") == 0) {
+			else if(dataType.compareTo("scalar") == 0) {
 				// scalar
 				dataId.setDataType(DataType.SCALAR);
 			}
@@ -1259,22 +1743,17 @@ public class DmlSyntacticValidator implements DmlListener
 			}
 			
 			valueType = paramCtx.paramType.valueType().getText();
-			if(valueType.compareTo("int") == 0 || valueType.compareTo("integer") == 0
-				|| valueType.compareTo("Int") == 0 || valueType.compareTo("Integer") == 0) {
+			if(valueType.compareTo("int") == 0) {
 				dataId.setValueType(ValueType.INT);
 			}
-			else if(valueType.compareTo("string") == 0 || valueType.compareTo("String") == 0) {
+			else if(valueType.compareTo("str") == 0) {
 				dataId.setValueType(ValueType.STRING);
 			}
-			else if(valueType.compareTo("boolean") == 0 || valueType.compareTo("Boolean") == 0) {
+			else if(valueType.compareTo("bool") == 0) {
 				dataId.setValueType(ValueType.BOOLEAN);
 			}
-			else if(valueType.compareTo("double") == 0 || valueType.compareTo("Double") == 0) {
+			else if(valueType.compareTo("float") == 0) {
 				dataId.setValueType(ValueType.DOUBLE);
-			}
-			else if(valueType.compareTo("bool") == 0) {
-				helper.notifyErrorListeners("invalid valuetype " + valueType + " (Quickfix: use \'boolean\' instead)", paramCtx.start);
-				return null;
 			}
 			else {
 				helper.notifyErrorListeners("invalid valuetype " + valueType, paramCtx.start);
@@ -1350,7 +1829,7 @@ public class DmlSyntacticValidator implements DmlListener
 				// Empty value allowed
 			}
 			else {
-				helper.notifyErrorListeners("the value of user parameter for external function should be of type string", ctx.start);
+				helper.notifyErrorListeners("the value of user parameter for external function should be of type str", ctx.start);
 				return;
 			}
 			otherParams.put(paramName, val);
@@ -1360,25 +1839,10 @@ public class DmlSyntacticValidator implements DmlListener
 		}
 		functionStmt.setOtherParams(otherParams);
 		if(!atleastOneClassName) {
-			helper.notifyErrorListeners("the parameter \'className\' needs to be passed for externalFunction", ctx.start);
+			helper.notifyErrorListeners("the parameter \'className\' needs to be passed for defExternal", ctx.start);
 			return;
 		}
-		
-//		if(ctx.body.size() > 0) {
-//			// handle function body
-//			// Create arraylist of one statement block
-//			ArrayList<StatementBlock> body = new ArrayList<StatementBlock>();
-//			for(StatementContext stmtCtx : ctx.body) {
-//				body.add(getStatementBlock(stmtCtx.info.stmt));
-//			}
-//			((ExternalFunctionStatement) functionStmt).setBody(body);
-//			((ExternalFunctionStatement) functionStmt).mergeStatementBlocks();
-//		}
-//		else {
-//			helper.notifyErrorListeners("functions with no statements are not allowed", ctx.start);
-//			return;
-//		}
-		
+				
 		ctx.info.stmt = functionStmt;
 		setFileLineColumn(ctx.info.stmt, ctx);
 		ctx.info.functionName = ctx.name.getText();
@@ -1465,11 +1929,20 @@ public class DmlSyntacticValidator implements DmlListener
 	@Override
 	public void exitMatrixDataTypeCheck(MatrixDataTypeCheckContext ctx) {
 		if(		ctx.ID().getText().compareTo("matrix") == 0 
-				|| ctx.ID().getText().compareTo("Matrix") == 0 
-				|| ctx.ID().getText().compareTo("Scalar") == 0
 				|| ctx.ID().getText().compareTo("scalar") == 0
 				) {
 			// Do nothing
+		}
+		else if(ctx.ID().getText().compareTo("Matrix") == 0)
+			helper.notifyErrorListeners("incorrect datatype (Hint: use matrix instead of Matrix)", ctx.start);
+		else if(ctx.ID().getText().compareTo("Scalar") == 0)
+			helper.notifyErrorListeners("incorrect datatype (Hint: use scalar instead of Scalar)", ctx.start);
+		else if(		ctx.ID().getText().compareTo("int") == 0 
+				|| ctx.ID().getText().compareTo("str") == 0
+				|| ctx.ID().getText().compareTo("bool") == 0
+				|| ctx.ID().getText().compareTo("float") == 0
+				) {
+			helper.notifyErrorListeners("expected datatype but found a valuetype (Hint: use matrix or scalar instead of " + ctx.ID().getText() + ")", ctx.start);
 		}
 		else {
 			helper.notifyErrorListeners("incorrect datatype (expected matrix or scalar)", ctx.start);
@@ -1496,8 +1969,8 @@ public class DmlSyntacticValidator implements DmlListener
 	
 	@Override
 	public void exitIterablePredicateSeqExpression(IterablePredicateSeqExpressionContext ctx) {
-		if(ctx.ID().getText().compareTo("seq") != 0) {
-			helper.notifyErrorListeners("incorrect function:\'" + ctx.ID().getText() + "\'. expected \'seq\'", ctx.start);
+		if(ctx.ID().getText().compareTo("range") != 0) {
+			helper.notifyErrorListeners("incorrect function:\'" + ctx.ID().getText() + "\'. expected \'range\'", ctx.start);
 			return;
 		}
 		ctx.info.from = ctx.from.info.expr;
@@ -1529,5 +2002,39 @@ public class DmlSyntacticValidator implements DmlListener
 		setFileLineColumn(ctx.info.expr, ctx);
 	}
 	
+	@Override
+	public void exitIgnoreNewLine(IgnoreNewLineContext ctx) {
+		// Introduce empty StatementInfo 
+		// This is later ignored by PyDMLParserWrapper
+		try {
+			ctx.info.stmt = new AssignmentStatement(null, null, 0, 0, 0, 0);
+			ctx.info.stmt.setEmptyNewLineStatement(true);
+		} catch (LanguageException e) {
+			e.printStackTrace();
+		}
 		
+	}
+	@Override
+	public void enterValueDataTypeCheck(ValueDataTypeCheckContext ctx) { }
+	@Override
+	public void exitValueDataTypeCheck(ValueDataTypeCheckContext ctx) {
+		if(		ctx.ID().getText().compareTo("int") == 0 
+				|| ctx.ID().getText().compareTo("str") == 0
+				|| ctx.ID().getText().compareTo("bool") == 0
+				|| ctx.ID().getText().compareTo("float") == 0
+				) {
+			// Do nothing
+		}
+		else if(ctx.ID().getText().compareTo("integer") == 0)
+			helper.notifyErrorListeners("incorrect valuetype (Hint: use int instead of integer)", ctx.start);
+		else if(ctx.ID().getText().compareTo("double") == 0)
+			helper.notifyErrorListeners("incorrect valuetype (Hint: use float instead of double)", ctx.start);
+		else if(ctx.ID().getText().compareTo("boolean") == 0)
+			helper.notifyErrorListeners("incorrect valuetype (Hint: use bool instead of boolean)", ctx.start);
+		else if(ctx.ID().getText().compareTo("string") == 0)
+			helper.notifyErrorListeners("incorrect valuetype (Hint: use str instead of string)", ctx.start);
+		else {
+			helper.notifyErrorListeners("incorrect valuetype (expected int, str, bool or float)", ctx.start);
+		}
+	}
 }
