+++ src/main/java/org/apache/sysml/parser/pydml/StatementInfo.java
@@ -17,20 +17,16 @@
  * under the License.
  */
 
-package org.apache.sysml.parser.antlr4;
+package org.apache.sysml.parser.pydml;
 
-import java.util.HashMap;
-import org.apache.sysml.parser.DMLProgram;
-
-public class StatementInfo {
-
-	public org.apache.sysml.parser.Statement stmt = null;
-	
-	// Valid only for import statements
-	public HashMap<String,DMLProgram> namespaces = null;
-	
-	// Valid only for function statement
-	//public String namespace = DMLProgram.DEFAULT_NAMESPACE;
-	public String functionName = "";
+/**
+ * This class exists solely to prevent compiler warnings.
+ * 
+ * <p>
+ * The ExpressionInfo and StatementInfo classes are shared among both parsers
+ * (R-like and Python-like dialects), and Antlr-generated code assumes that
+ * these classes are present in the parser's namespace.
+ */
+class StatementInfo extends org.apache.sysml.parser.dml.StatementInfo {
 
 }
